requirements.txt
cocotb
cocotbext-axi
cocotbext-uart
pytest
